+++ IPython/core/debugger.py
@@ -132,7 +132,6 @@ import warnings
 from contextlib import contextmanager
 from functools import lru_cache
 
-from IPython import get_ipython
 from IPython.core.debugger_backport import PdbClosureBackport
 from IPython.utils import PyColorize
 from IPython.utils.PyColorize import TokenStream
@@ -243,6 +242,8 @@ class Pdb(OldPdb):
         stdin=None,
         stdout=None,
         context: int | None | str = 5,
+        *,
+        shell: TerminalInteractiveShell = None,
         **kwargs,
     ):
         """Create a new IPython debugger.
@@ -258,6 +259,7 @@ class Pdb(OldPdb):
         context : int
             Number of lines of source code context to show when
             displaying stacktrace information.
+        shell : InteractiveShell
         **kwargs
             Passed to pdb.Pdb.
 
@@ -282,16 +284,6 @@ class Pdb(OldPdb):
 
         # IPython changes...
         self.shell = get_ipython()
-
-        if self.shell is None:
-            save_main = sys.modules["__main__"]
-            # No IPython instance running, we must create one
-            from IPython.terminal.interactiveshell import TerminalInteractiveShell
-
-            self.shell = TerminalInteractiveShell.instance()
-            # needed by any code which calls __import__("__main__") after
-            # the debugger was entered. See also #9941.
-            sys.modules["__main__"] = save_main
 
         self.aliases = {}
