+++ IPython/core/profiledir.py
@@ -1,4 +1,3 @@
-# encoding: utf-8
 """An object for managing IPython profile directories."""
 
 # Copyright (c) IPython Development Team.
@@ -7,10 +6,10 @@
 import os
 import shutil
 import errno
+import IPython
 from pathlib import Path
 
 from traitlets.config.configurable import LoggingConfigurable
-from ..paths import get_ipython_package_dir
 from ..utils.path import expand_path, ensure_dir_exists
 from traitlets import Unicode, Bool, observe
 
@@ -242,3 +241,10 @@ class ProfileDir(LoggingConfigurable):
         if not os.path.isdir(profile_dir):
             raise ProfileDirError('Profile directory not found: %s' % profile_dir)
         return cls(location=profile_dir, config=config)
+
+
+def get_ipython_package_dir() -> str:
+    """Get the base directory where IPython itself is installed."""
+    ipdir = os.path.dirname(IPython.__file__)
+    assert isinstance(ipdir, str)
+    return ipdir
