+++ IPython/terminal/debugger.py
@@ -2,8 +2,10 @@ import asyncio
 import os
 import sys
 
+from IPython import get_ipython
 from IPython.core.debugger import Pdb
 from IPython.core.completer import IPCompleter
+from IPython.terminal.interactiveshell import TerminalInteractiveShell
 from .ptutils import IPythonPTCompleter
 from .shortcuts import create_ipython_shortcuts
 from . import embed
@@ -26,7 +28,16 @@ class TerminalPdb(Pdb):
     """Standalone IPython debugger."""
 
     def __init__(self, *args, pt_session_options=None, **kwargs):
-        Pdb.__init__(self, *args, **kwargs)
+        shell = get_ipython()
+        if shell is None:
+            save_main = sys.modules["__main__"]
+            # No IPython instance running, we must create one
+            shell = TerminalInteractiveShell.instance()
+            # needed by any code which calls __import__("__main__") after
+            # the debugger was entered. See also #9941.
+            sys.modules["__main__"] = save_main
+
+        Pdb.__init__(self, *args, shell=shell, **kwargs)
         self._ptcomp = None
         self.pt_init(pt_session_options)
         self.thread_executor = ThreadPoolExecutor(1)
