+++ IPython/terminal/interactiveshell.py
@@ -46,7 +46,6 @@ from prompt_toolkit.styles.pygments import style_from_pygments_cls, style_from_p
 from pygments.styles import get_style_by_name
 from pygments.style import Style
 
-from .debugger import TerminalPdb, Pdb
 from .magics import TerminalMagics
 from .pt_inputhooks import get_inputhook_name_and_func
 from .prompts import Prompts, ClassicPrompts, RichPromptDisplayHook
@@ -246,6 +245,7 @@ class TerminalInteractiveShell(InteractiveShell):
 
     @property
     def debugger_cls(self):
+        from .debugger import TerminalPdb, Pdb
         return Pdb if self.simple_prompt else TerminalPdb
 
     confirm_exit = Bool(True,
