+++ IPython/utils/io.py
@@ -15,10 +15,6 @@ import tempfile
 from pathlib import Path
 from warnings import warn
 
-from IPython.utils.decorators import undoc
-from .capture import CapturedIO, capture_output
-
-
 class Tee:
     """A class to duplicate an output stream to stdout/err.
